更新了一周心情接口

+++ src/main/java/com/testdb/demo/entity/Mood.java
@@ -14,7 +14,7 @@ public class Mood {
 
     @TableId(type = IdType.AUTO)
     private String id;
-    private long moodType;
+    private int moodType;
     private String description;
     private LocalDate moodDate;
     private String dayOfWeek;
